requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
django-elasticsearch-dsl==8.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
elastic-transport==8.12.0
elasticsearch==8.12.1
elasticsearch-dsl==8.12.0
idna==3.6
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tokenauth==0.9.1
typing_extensions==4.10.0
urllib3==2.2.1
